[crankshaft] [ia32] Remove dynamic frame alignment optimization.
This optimization does not give us much (see perf try bot results associated with this CL) but complicates things a lot. The main motivation is to avoid additional complexity in tail call optimization.

There are some pieces left in the deoptimizer, but I'll address this in a separate CL.

Review URL: https://codereview.chromium.org/1731273003

Cr-Commit-Position: refs/heads/master@{#34273}